Corrected response handling in example code
This API endpoint returns an object, but the `then` handler was expecting an array. I changed the parameter to represent the response from the API, and then access the expected list on the `data` property. 

I did not test this locally, but this change to the example code works when I tested it in the editor used by the docs.